+++ test/dart_test.dart
@@ -9,34 +9,15 @@ void main() {
 
   group('Pixelmatch Tests', () {
     diffTest('1a', '1b', '1diff', options, 143);
-    diffTest('1a', '1b', '1diffmask',
-        {'threshold': 0.05, 'includeAA': false, 'diffMask': true}, 143);
-    diffTest(
-        '1a', '1a', '1emptydiffmask', {'threshold': 0, 'diffMask': true}, 0);
-    diffTest(
-        '2a',
-        '2b',
-        '2diff',
-        {
-          'threshold': 0.05,
-          'alpha': 0.5,
-          'aaColor': [0, 192, 0],
-          'diffColor': [255, 0, 255]
-        },
-        12437);
+    diffTest('1a', '1b', '1diffmask', {'threshold': 0.05, 'includeAA': false, 'diffMask': true}, 143);
+    diffTest('1a', '1a', '1emptydiffmask', {'threshold': 0, 'diffMask': true}, 0);
+    diffTest('2a', '2b', '2diff', {'threshold': 0.05, 'alpha': 0.5, 'aaColor': [0, 192, 0], 'diffColor': [255, 0, 255]}, 12437);
     diffTest('3a', '3b', '3diff', options, 212);
     diffTest('4a', '4b', '4diff', options, 36049);
     diffTest('5a', '5b', '5diff', options, 0);
     diffTest('6a', '6b', '6diff', options, 51);
     diffTest('6a', '6a', '6empty', {'threshold': 0}, 0);
-    diffTest(
-        '7a',
-        '7b',
-        '7diff',
-        {
-          'diffColorAlt': [0, 255, 0]
-        },
-        2448);
+    diffTest('7a', '7b', '7diff', {'diffColorAlt': [0, 255, 0]},2448);
   });
 }
 
@@ -58,10 +39,8 @@ void diffTest(String imgPath1, String imgPath2, String diffPath,
         img1.toUint8List(), img2.toUint8List(), null, width, height, options);
 
     expect(diff.toUint8List(), equals(out), reason: 'diff is different');
-    expect(mismatch, equals(expectedMismatch),
-        reason: 'number of mismatched pixels');
-    expect(mismatch, equals(mismatch2),
-        reason: 'number of mismatched pixels without diff');
+    expect(mismatch, equals(expectedMismatch), reason: 'number of mismatched pixels');
+    expect(mismatch, equals(mismatch2), reason: 'number of mismatched pixels without diff');
   });
 }
 
@@ -70,7 +49,3 @@ Image readImage(String name) {
   final image = decodePng(bytes);
   return image!.convert(numChannels: 4);
 }
-
-// void writeImage(String name, Image image) {
-//   File('test/$name.png').writeAsBytesSync(encodePng(image));
-// }