+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ## 1.0.0
 
-- Initial version
+- Initial ported version
 
 ## 1.0.1
 
 - Fix sdk dependency
+
+## 1.1.0
+
+- Utils for integration_testing
+- Universal flutter and dart sdk package